Validator::formatMessage: support dynamic error messages
This is useful when you want to show user an informative error message, whose text
depends on some external data that depends on the value of some other field.

For example, my form allows user to choose a category of a team. Each category has a different
limit for number of allowed members, which is specified in config.neon. I want to show the user
the limit as part of the error message when she manages to exceed the limit.