DefaultFormRenderer: Fix GET Application\UI\Form
Forms\Form allows to pass stringable objects to setAction method,
which is also used by Application\UI\Form to store Application\UI\Link object.
DefaultFormRenderer calls string manipulation functions on the Form’s action
when rendering, causing it to raise a TypeError when strict types are enabled.

To fix this, I am stringifying the action in the DefaultFormRenderer.
I have opted to do it there, rather than in the Form::setAction method,
in case the Link object is not ready yet at the time the Form is created.